test(runs): move lab bundle regression out of command file
